+++ client/package.json
@@ -42,7 +42,8 @@
     "tls-browserify": "^0.2.2",
     "url": "^0.11.0",
     "util": "^0.12.5",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "react-scripts start",
